requirements.txt
appdirs==1.4.0
chainer==1.0.1
Cython==0.22.1
decorator==3.4.2
h5py==2.5.0
Mako==1.0.1
MarkupSafe==0.23
nose==1.3.7
numpy==1.9.2
py==1.4.30
pycuda==2015.1
pytest==2.7.2
python-shogi==0.0.1
pytools==2014.3.5
scikit-learn==0.16.1
scikits.cuda==0.5.0b2
scipy==0.15.1
six==1.9.0
